# Conflicts:
#	src/fineman/finemans_algorithm.py
#	src/scripts/random_graph_no_neg_cycles_gen.py
#	src/scripts/synthetic_graph_generator.py
#	src/tests/test_finemans_algorithm.py
#	src/weight_type.py

+++ src/fineman/finemans_algorithm.py
@@ -6,7 +6,7 @@ from src.fineman.dijkstra import dijkstra
 from src.fineman.elimination_algorithm import elimination_algorithm
 from src.weight_type import WEIGHT_TYPE
 
-def fineman(graph, source: int, seed = None, type = int):
+def fineman(graph, source: int, seed = None, type = float):
     WEIGHT_TYPE = type
     if seed is not None: rand.seed(seed)
