+++ src/weight_type.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 # This variable determines the type of the weights of the algorithm
 
-WEIGHT_TYPE = int
+WEIGHT_TYPE = float
 
 types = {
         "decimal": Decimal,
